Создан модуль data.js. Рефакторинг кода
